requirements.txt
charset-normalizer==3.3.2
djangorestframework==3.14.0
docker==6.1.3
packaging==23.2
requests==2.31.0
websocket-client==1.6.4
Django~=3.2.12
